front_app/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
jdatetime==4.1.1
pytz==2023.3